feat!: Updated command folder structure for new command handler
