+++ src/commands/info/help.js
