+++ src/commands/player/loop.js
