+++ src/commands/player/nowplaying.js
