+++ src/commands/player/pause.js
