+++ src/commands/player/play.js
