+++ src/commands/player/remove.js
